refactor(mobile): move contactForm to a container
